* b2Alloc now takes a std::size_t argument instead of a signed int * the original b2Alloc and b2Free implementations in b2Settings have been moved to Snapbox2D/storage * fixed the install path of the header files * adding initial SnapBox2D/storage implementation

+++ Box2D/Box2D/Common/b2Settings.h
@@ -128,7 +128,7 @@ typedef double float64;
 // Memory Allocation
 
 /// Implement this function to use your own memory allocator.
-void* b2Alloc(int32 size);
+void* b2Alloc(size_t size);
 
 /// If you implement b2Alloc, you should also implement this function.
 void b2Free(void* mem);
